Change lst function to take into account name changes
The user can first run the pipeline, with a given threshold on L90 and nb contigs. Then, if he wants to redo it, changing the threshold (and then the strain numbers of the genomes, and hence their name), it is not necessary to run prokka again with the new genome names. Instead, when reading the tbl file, replace the genome name by the current one.