CI: unit and functional tests
We now have 2 separated steps for unit-tests and functional-tests. The 1st advantage is that they can run independently, and hence, in parallel.
After, we combine their covers to generate the html page.

We could not run unit and functional tests with the same run (py.test test) because logs are not handled, and hence, tests on error messages fail.