Change input file format, name and/or date ok
Format of input file changed, to allow the user to give a specific name and/or date to a genome.
There is still one line per genome, first word is the genome sequence file, and then, it is separated from the other informations by '::'.
The other informations (after '::') can be name.date or name or name. or .date