Change behaviour of run_prokka with -F option
Before, -F option added the '--force' argument to prokka command line. However, this option does not remove the existing files in prokka dir. It only overrides the files which arre recreated with the same name. So, if prokka ran for a genome, with a given output name, and now has to run again (-F option), but with another output name, both result files will be present in prokka dir (those with previous output name, and those with new output name). While checking the prokka output dir, there will be several tbl/ffn/faa files, and we will consider that there was a problem running prokka.
Now, if -F: remove prokka directory. It will create a new one, with only the new files.

-F -> force=True (instead of '--force' before); default: force=False