Don't call barrnap -h when it is not supposed to be installed.
We called barrnap -h at the end of each install job, to check that it was installed hen it should be, and absent when it shouldn't.
However, when it is not installed, barrnap -h return code is not null, and the job fails. So, now, when it is not installed,
call barrnap -h through 'echo' to avoid error.