Update functional tests
Before, when analysing genomes, we only created a new file if split5N was on. Now, we always create a new file, with short contig names, to directly give it to prokka. We don't need to, before running prokka, change the header names to 'gembase' format as we did before.