Correct bug renaming headers
While renaming contig headers when cutting by each stretch of 5N, we restarted at number 0 for each new contig (and then 1, 2 etc. for each of its cut subparts). But, as we keep only the 15 first characters of the original contig name, if those characters are the same, then several contigs can have the same name. So, we give unique contig IDs.