Use a QueueHandler for logs inside subprocesses
Before, inside subprocess, logs where done on the same logger, with no caution about multiple access to a same file...
Now, inside subprocesses, logs are given to a QueueHandler. A thread is listening to the Queue, and handling each record to give it to the main logger: only 1 process is writing into log files.