Do not re-do alignment if file already exists
if align file exists and is ok -> go to back-translate step. Otherwise, remove it as well as back-translate file
if back-translate exists and is ok -> go to add_missing step. Otherwise, remove it and redo back-translate step
if back-translate is already ok (no missing genomes, or done in a previous step) -> go to next step. Otherwise, add missing genomes