npz matrix not re-calculated
In 'compare_all', we checked is the txt matrix was already here. If yes, we read matrix from this file.
But we did not check if npz matrix was already here! We assumed that if there is not txt, then there is no txt
Now, does not redo calculations if either npz or txt exist