test all: bigger L90 for g3 and g4
for functional tests for 'all' module, we use 4 genomes: g1, g2, g3, g4
In "test_main_norefseq", we annotate them with prodigal. It trains on the first genome, which is the one with best quality. For now, g1, g3 and g4 all had L90=1 and 1 contig, so, there order was random. But prodigal can only run on g1. The 2 others are too small. So, we increase nbcont of g3 and g4 to be sure that it will always train on g1.